requirements.txt
# Core
numpy>=1.21
pandas>=1.3
matplotlib>=3.4
seaborn>=0.11
scikit-learn>=0.24
xgboost>=1.5
joblib>=1.2
tqdm>=4.60
pandarallel>=1.6
transformers>=4.30

# Deep Learning
tensorflow>=2.10
torch>=1.10
pytorch-tabnet>=3.1

# Data Synthesis
sdv>=1.4

# NLP / Embeddings / RAG
langchain>=0.1.17
langchain-community>=0.0.30
langchain-google-genai>=0.0.8
faiss-cpu>=1.7

# Web & UI
streamlit>=1.10
gradio>=3.40
Pillow>=8.0

# Voice & Audio
whisper @ git+https://github.com/openai/whisper.git
pyttsx3>=2.90
scipy>=1.7

# Data Access
kagglehub>=0.1.6
